Use InstructionBytes for instructions, add README
This refactors instructions to include the bytes parsed when executing
the instruction rather than trying to read from PC and find the few
bytes. This has the advantage in making it easier to print the output of
the instructions as well as knowing the offset to the next instruction.

This also fixes branching and jump issues and simplifies the get_operand
and get_operand_address into a single function.

Finally this adds a README.md which has some dependencies spelled out.